switcht to int64 for indexing

+++ edge_buffer.hpp
@@ -6,7 +6,7 @@
 #include <tskit.h>
 #include "tskit_tools.hpp"
 
-using EDGE_BUFFER_INDEX_TYPE = std::int32_t;
+using EDGE_BUFFER_INDEX_TYPE = std::int64_t;
 static const EDGE_BUFFER_INDEX_TYPE NULL_EDGE_BUFFER_INDEX = -1;
 
 struct BirthData
